+++ src/chatevent/__init__.py
@@ -34,4 +34,4 @@ __all__ = [
     "temporary_zulip_topic_watch",
     "zulip_topic_target",
 ]
-__version__ = "0.2.4"
+__version__ = "0.2.5"